clip_model/requirements.txt
# requirements.txt for CLIP model service
torch
transformers
Pillow
fastapi
uvicorn
python-multipart
opencv-python-headless
numpy
